+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## UPCOMING
 **_Add new changes here as they land_**
 
-- Avoid dev-mode console error with React 18 when using shared selectors across multiple `<RecoilRoot>`'s. (#1810)
+- Avoid dev-mode console error with React 18 when using shared async selectors across multiple `<RecoilRoot>`'s. (#1712)
 - Update typing for family parameters to better support Map, Set, and classes with `toJSON()`. (#1709, #1703)
+- Cleanup potential memory leak when using async selectors (#1714)
+- Fix potentially hung async selectors when shared across multiple roots that depend on atoms initialized with promises that don't resolve (#1714)
 
 ## 0.7 (2022-03-31)